Split editing server/host code out to new package "sitecore-jss-nextjs-editing-host"
