+++ packages/sitecore-jss-nextjs-editing-host/.gitignore
@@ -1,0 +1,1 @@
+types/
