+++ packages/sitecore-jss-nextjs-editing-host/.npmignore
