+++ packages/sitecore-jss-nextjs-editing-host/README.md
@@ -1,0 +1,6 @@
+# Sitecore JavaScript Services Editing Host for Next.js
+
+This module is provided as a part of Sitecore JavaScript Services. It contains the editing host implementation for Next.js.
+
+Consult the [Sitecore JSS documentation](https://jss.sitecore.net) for further details.
+
