+++ packages/sitecore-jss-nextjs-editing-host/package.json
@@ -1,14 +1,12 @@
 {
-  "name": "@sitecore-jss/sitecore-jss-nextjs",
+  "name": "@sitecore-jss/sitecore-jss-nextjs-editing-host",
   "version": "15.0.1",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "sideEffects": false,
+  "main": "dist/index.js",
   "scripts": {
-    "build": "npm run clean && tsc -p tsconfig.json && tsc -p tsconfig-esm.json",
+    "build": "npm run clean && tsc",
     "clean": "del-cli dist types",
     "lint": "tslint -p . -c ./tslint.json \"src/**/*.ts\"",
-    "test": "mocha --require ts-node/register --require ./src/tests/shim.ts ./src/tests/jsdom-setup.ts ./src/tests/enzyme-setup.ts \"./src/**/*.test.ts\" \"./src/**/*.test.tsx\"",
+    "test": "mocha --require ts-node/register \"./src/**/*.test.ts\"",
     "prepublishOnly": "npm run build",
     "coverage": "nyc --require ts-node/register npm test"
   },
@@ -25,40 +23,34 @@
     "url": "https://github.com/sitecore/jss/issues"
   },
   "devDependencies": {
+    "@sitecore-jss/sitecore-jss-nextjs": "^15.0.1",
+    "@types/body-parser": "^1.19.0",
     "@types/chai": "^4.2.2",
-    "@types/chai-spies": "^1.0.1",
-    "@types/chai-string": "^1.4.2",
-    "@types/enzyme": "^3.10.3",
-    "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/compression": "^1.7.0",
+    "@types/express": "^4.17.9",
     "@types/mocha": "^5.2.7",
     "@types/node": "12.7.5",
-    "@types/react": "16.9.2",
-    "@types/react-dom": "^16.9.0",
     "chai": "^4.2.0",
-    "chai-spies": "^1.0.0",
-    "chai-string": "^1.5.0",
+    "chalk": "^4.1.0",
     "del-cli": "^3.0.1",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "jsdom": "^15.1.1",
     "mocha": "^8.1.3",
     "next": "^10.0.3",
     "nyc": "^15.1.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "react-test-renderer": "^16.9.0",
     "ts-node": "^9.0.0",
-    "tslint-react": "^4.0.0",
     "typescript": "^4.0.2"
   },
   "peerDependencies": {
+    "@sitecore-jss/sitecore-jss-nextjs": "^15.0.1",
     "next": "^10.0.3",
     "react": "^16.9.0",
     "react-dom": "^16.9.0"
   },
   "dependencies": {
-    "@sitecore-jss/sitecore-jss": "^15.0.1",
-    "@sitecore-jss/sitecore-jss-react": "^15.0.1"
+    "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
+    "express": "^4.17.1"
   },
   "description": "",
   "types": "types/index.d.ts",
