+++ packages/sitecore-jss-nextjs-editing-host/src/absolutify.ts
