+++ packages/sitecore-jss-nextjs-editing-host/src/index.ts
@@ -1,0 +1,4 @@
+export { 
+  EditingServerOptions,
+  startEditingServer
+} from './editing-server';
