+++ packages/sitecore-jss-nextjs/src/sharedTypes/editing-data.ts
