+++ packages/sitecore-jss-nextjs/tsconfig-esm.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "module": "CommonJS",
-    "outDir": "dist/cjs"
+    "module": "es6",
+    "outDir": "dist/esm"
   },
 }
